% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{epoch}
\alias{epoch}
\title{Performs an epoch.}
\usage{
epoch(X, Y, w, b, eta)
}
\arguments{
\item{X}{All inputs in the training set.}

\item{Y}{All expected outputs in the training set.}

\item{w}{Weights of NN.}

\item{b}{Biases of NN.}

\item{eta}{Learning rate.}
}
\description{
Performs an epoch by running \code{single_step} for each pair from \code{X} and \code{Y}
}
