% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{optimize}
\alias{optimize}
\title{Runs gradient descent for a number of epochs.}
\usage{
optimize(X, Y, w, b, eta, epochs)
}
\arguments{
\item{X}{All inputs in the training set.}

\item{Y}{All expected outputs in the training set.}

\item{w}{Weights of NN.}

\item{b}{Biases of NN.}

\item{eta}{Learning rate.}

\item{epochs}{Number of epochs to run gradient descent for.}
}
\value{
\code{tibble} with weights, biases and their respective loss at the end of each epoch.
}
\description{
Runs gradient descent for a number of epochs.
}
