% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{reporting}
\alias{reporting}
\alias{data_response}
\alias{data_neurons}
\title{Reporting}
\usage{
data_response(X, Y, rsp, w, b, epoch)

data_neurons(X, w, b, epoch)
}
\arguments{
\item{X}{Vector of inputs to NN.}

\item{w}{NN weights}

\item{b}{NN biases}

\item{epoch}{A row from the optimization output to extract weights and biases from.}
}
\description{
Provide \code{w} and \code{b} or \code{epoch} but not all at the same time.
}
