% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{single_step}
\alias{single_step}
\title{Performs a single foward+backward step.}
\usage{
single_step(x, y, w, b, eta)
}
\arguments{
\item{x}{Scalar input.}

\item{y}{Expected scalar output.}

\item{w}{Weights of NN.}

\item{b}{Biases of NN.}

\item{eta}{Learning rate.}
}
\description{
Computes the response of the NN, all necessary derivaties and
updates parameters \code{w} and \code{b} according to the learning rate \code{eta}.
}
